LayoutTests/accessibility/mac/text-input-marked-text-marker-range-expected.txt

This test ensures that the input method marked range is available to accessibility clients as text marker range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !text is true
PASS text is 'test'
PASS !text is true
PASS element.stringValue is 'AXValue: test'
PASS text is ' message'
PASS element.stringValue is 'AXValue: test message'
PASS successfullyParsed is true

TEST COMPLETE
